requirements.txt
allure-pytest
coverage
pytest
pytest-cov
